Feature/issues202005 (#15)
* Issue #108

* Issue #111

* Issue #108

* Issue #109

* Issue #109

* Issue #112

* Issue #112

* Issue #112

* Issue #107

* Issue #113

* Issue #114

* issue$104

* Issue #105

* Issue #105

* Issue #115

* Issue #117

* Issue #116

* Issue #115

* Issue #126

* Issue #126

* Issue #126

* issue#126

* Issue #130

* Issue #130

* 動画視聴対応

* Change version to 1.5.0

* Issue #138

* Issue #136

* issue#126(commitミス対応)

* Issue #135

* Issue #140

* Issue #128

* Issue #140

* Issue #140

* Issue #118

* Issue #135

* Issue #122

* Issue #129

* Issue  #141

* Issue #140

* Issue #135

* Issue #143

* Issue #144

* Issue #144

* Issue #147

* Issue #144

* Issue #142

* Issue #148

* Issue #145

* Issue #141

* Issue #141

* Issue #141

* Issue #123

* Issue #135

* Issue #134

+++ app/models/help.rb
@@ -63,17 +63,21 @@ class Help < ApplicationRecord
 
   def create_file
     FileUtils.mkdir_p(get_help_path)
-    upload_file = get_help_path + self.link_name
+    upload_file = get_help_path + self.file_name
     File.open(upload_file, "wb") {|f|f.write(file.read)}
   end
 
   def check_exist_file
-    file_path = get_help_path + self.link_name
-    if File.exist?(file_path)
+    file_path = get_help_path + self.file_name
+    if self.file_name.present? && File.exist?(file_path)
 			## 既存のファイルがある場合はリネームする
       new_file_dir = get_help_path + "bak/"
       check_user_path(new_file_dir)
-      new_file_path = new_file_dir + self.file_name_was
+      if self.file_name_in_database.blank?
+        new_file_path = new_file_dir + self.file_name
+      else
+        new_file_path = new_file_dir + self.file_name_in_database
+      end
 			false unless FileUtils.move(file_path, new_file_path)
 		end
 
@@ -81,7 +85,7 @@ class Help < ApplicationRecord
   end
 
   def get_link_url
-    DOWNLOAD_PATH + HELP_PATH + self.link_name
+    DOWNLOAD_PATH + HELP_PATH + self.file_name
   end
 
   def get_help_path
